+++ howard/functions/commons.py
@@ -120,6 +120,9 @@ DEFAULT_DBSNP_URL = "https://ftp.ncbi.nih.gov/snp/archive"
 # Databases default folder
 DEFAULT_DATABASE_FOLDER = os.path.join(folder_howard_home, "databases")
 DEFAULT_ANNOTATIONS_FOLDER = f"{DEFAULT_DATABASE_FOLDER}/annotations/current"
+DEFAULT_PARQUET_FOLDER = f"{DEFAULT_DATABASE_FOLDER}/parquet/current"
+DEFAULT_BCFTOOLS_FOLDER = f"{DEFAULT_DATABASE_FOLDER}/bcftools/current"
+DEFAULT_BIGWIG_FOLDER = f"{DEFAULT_DATABASE_FOLDER}/bigwig/current"
 DEFAULT_GENOME_FOLDER = f"{DEFAULT_DATABASE_FOLDER}/genomes/current"
 DEFAULT_SNPEFF_FOLDER = f"{DEFAULT_DATABASE_FOLDER}/snpeff/current"
 DEFAULT_ANNOVAR_FOLDER = f"{DEFAULT_DATABASE_FOLDER}/annovar/current"
@@ -629,18 +632,23 @@ def find(name: str, path: str) -> str:
     return ""
 
 
-def find_all(name: str, path: str) -> list:
+def find_all(name: str, path: str, allow_dir: bool = True) -> list:
     """
-    "Walk the directory tree starting at path, and for each regular file with the name name, append its
-    full path to the result list."
+    "Walk the directory tree starting at path, and for each regular file or directory with the name name,
+    append its full path to the result list."
 
     The os.walk function is a generator that yields a 3-tuple containing the name of a directory, a list
     of its subdirectories, and a list of the files in that directory. The name of the directory is a
     string, and the lists of subdirectories and files are lists of strings
 
     :param name: The name of the file you're looking for
+    :type name: str
     :param path: The path to search in
-    :return: A list of all the files in the directory that have the name "name"
+    :type path: str
+    :param allow_dir: If True, the function will also search for directories with the specified name,
+    :type allow_dir: bool
+    :return: A list of all the files or directories in the directory that have the name "name"
+    :rtype: list
     """
 
     # result
@@ -653,6 +661,10 @@ def find_all(name: str, path: str) -> list:
         for filename in fnmatch.filter(files, name):
             if os.path.exists(os.path.join(root, filename)):
                 result.append(os.path.join(root, filename))
+        if allow_dir:
+            for dirname in fnmatch.filter(dirs, name):
+                if os.path.exists(os.path.join(root, dirname)):
+                    result.append(os.path.join(root, dirname))
     return result
 
 
@@ -2622,7 +2634,30 @@ def concat_and_compress_files(
             # Remove tmp file
             os.remove(output_file_tmp)
         except:
-            raise ValueError(f"Output file sorting failed: {output_file_tmp}")
+            log.warning(
+                f"Output file sorting (without tabix) failed: {output_file_tmp}"
+            )
+            # Sort with pysam with index before
+            try:
+                pysam.tabix_index(output_file_tmp, preset="vcf", force=True)
+                pysam.bcftools.sort(
+                    f"-Oz{compression_level}",
+                    "-o",
+                    output_file,
+                    "-T",
+                    output_file_tmp,
+                    output_file_tmp,
+                    "-m",
+                    f"{memory}G",
+                    threads=threads,
+                    catch_stdout=False,
+                )
+                # Remove tmp file
+                os.remove(output_file_tmp)
+            except:
+                raise ValueError(
+                    f"Output file sorting (with tabix) failed: {output_file_tmp}"
+                )
 
     else:
         # Rename tmp file
@@ -4800,7 +4835,7 @@ def convert_markdown_to_pdf(input_file: str, output_file: str) -> None:
     convert_html_to_pdf(temp_html_path, output_file)
 
 
-def cast_columns_query(query, conn):
+def cast_columns_query(query, conn, sep: str = ","):
     """
     Cast columns of a query to VARCHAR or aggregate arrays to strings.
 
@@ -4808,6 +4843,8 @@ def cast_columns_query(query, conn):
     :type query: str
     :param conn: The database connection.
     :type conn: duckDB connection
+    :param sep: The separator for list aggregation, defaults to ","
+    :type sep: str
     :return: The modified SQL query with casted columns.
     :rtype: str
     """
@@ -4826,7 +4863,7 @@ def cast_columns_query(query, conn):
         column_type = description_dict.get(column, {}).get("type", "VARCHAR")
         if column_type.endswith("[]"):
             list_columns.append(
-                f""" list_aggregate("{column}", 'string_agg', ',') AS '{column}' """
+                f""" list_aggregate("{column}", 'string_agg', '{sep}') AS '{column}' """
             )
         else:
             list_columns.append(f""" "{column}" """)
@@ -4838,3 +4875,92 @@ def cast_columns_query(query, conn):
     """
 
     return query_cast
+
+
+def annotation_file_find(
+    annotation_file: str = None,
+    databases_folders: list = [],
+    assembly: str = None,
+) -> str:
+    """
+    The function `annotation_file_find` searches for a specified annotation file in a list of
+    directories, including assembly-specific folders, and returns the full path of the found file.
+    :param annotation_file: The `annotation_file` parameter is a string that represents the name of the
+    annotation file you are looking for. This file may be located in different directories or folders
+    :type annotation_file: str
+    :param databases_folders: The `databases_folders` parameter is a list of directories or folders
+    where the function will search for the specified annotation file. These directories may contain
+    different versions or assemblies of the annotation file
+    :type databases_folders: list
+    :param assembly: The `assembly` parameter is a string that represents the specific assembly or
+    version of the annotation file you are looking for. It is used to narrow down the search to a
+    specific assembly folder within the provided list of `databases_folders`
+    :type assembly: str (optional)
+    :return: The function `annotation_file_find` returns the full path of the found annotation file as a
+    string. If the file is not found in any of the specified directories, it returns `None`.
+    If the file is found, it returns the full path to the file.
+    If the file is not found, it returns None.
+    :rtype: str or None
+    """
+
+    # Init
+    annotation_file_found = None
+
+    # Check if annotation_file exists
+    if os.path.exists(annotation_file):
+        annotation_file_found = annotation_file
+
+    # Check if full annotation_file exists
+    elif os.path.exists(full_path(annotation_file)):
+        annotation_file_found = full_path(annotation_file)
+
+    else:
+
+        # Find within assembly folders
+        if assembly is not None:
+            assembly_folders = [
+                full_path(os.path.join(database, assembly))
+                for database in databases_folders
+            ]
+            log.debug(
+                f"Searching for {annotation_file} in assembly folders: {assembly_folders}"
+            )
+            found_file = search_in_folders(annotation_file, assembly_folders)
+            log.debug(f"Found file in assembly folders: {found_file}")
+            if found_file:
+                return found_file
+
+        # Search in general folders
+        general_folders = [
+            full_path(os.path.join(database, os.path.dirname(annotation_file)))
+            for database in databases_folders
+        ]
+        return search_in_folders(os.path.basename(annotation_file), general_folders)
+
+    return annotation_file_found
+
+
+def search_in_folders(file_name: str, search_folders: list) -> str:
+    """
+    The function `search_in_folders` searches for a specified file in a list of directories and returns
+    the full path of the first found file.
+    :param file_name: The `file_name` parameter is a string that represents the name of the file you are
+    looking for. This file may be located in different directories or folders
+    :type file_name: str
+    :param search_folders: The `search_folders` parameter is a list of directories or folders where the
+    function will search for the specified file. These directories may contain different versions or
+    assemblies of the file
+    :type search_folders: list
+    :return: The function `search_in_folders` returns the full path of the first found file as a string.
+    If the file is not found in any of the specified directories, it returns `None`.
+    :rtype: str or None
+    """
+
+    for folder in search_folders:
+        log.debug(f"Searching for {file_name} in folder: {folder}")
+        found_files = find_all(file_name, full_path(folder))
+        log.debug(f"Found files: {found_files}")
+        if found_files:
+            return found_files[0]
+
+    return None