+++ howard/objects/database.py
@@ -1772,8 +1772,50 @@ class Database:
             table = self.get_database_table(database=database)
 
         if sql_query:
-            columns_list = list(database.query(sql_query).columns)
-            return columns_list
+
+            columns_list = None
+
+            # Add robust error catching
+            try:
+                limited_query = f"SELECT * FROM ({sql_query}) AS _limited_query LIMIT 1"
+                if type(database) == duckdb.DuckDBPyConnection:
+                    # If the database is a DuckDB connection, execute the query directly
+                    try:
+                        columns_list = list(database.query(limited_query).columns)
+                        log.debug(
+                            f"Successfully executed query with database connection"
+                        )
+                    except Exception as e:
+                        log.debug(
+                            f"Error executing SQL query with database connection: {e}"
+                        )
+                        columns_list = None
+
+                # If no result or if database is not a DuckDB connection
+                else:  # if columns_list is None:
+                    try:
+                        limited_query = (
+                            f"SELECT * FROM ({sql_query}) AS _limited_query LIMIT 1"
+                        )
+                        q = self.query(limited_query)
+                        columns_list = list(q.columns)
+                        log.debug(f"Successfully executed query with self.conn")
+                    except Exception as e:
+                        log.debug(f"Error executing SQL query with self.conn: {e}")
+                        columns_list = None
+
+            except Exception as e:
+                log.debug(f"Unexpected error during query execution: {e}")
+                columns_list = None
+
+            # If we have columns, return them
+            if columns_list:
+                return columns_list
+
+        # if sql_query and type(database) == duckdb.DuckDBPyConnection:
+        #     columns_list = list(database.query(sql_query).columns)
+        #     log.debug(f"Columns from SQL query: {columns_list}")
+        #     return columns_list
 
         try:
             if database and self.exists(database):
@@ -1815,10 +1857,19 @@ class Database:
                     sql_query = f"SELECT * FROM {sql_from} LIMIT 0"
 
                     # Get columns
-                    result_description = self.conn.execute(sql_query).description
+                    # result_description = self.conn.execute(sql_query).description
+                    result_columns = self.conn.query(sql_query).columns
 
-                    # Extract columns' names
-                    columns = [desc[0] for desc in result_description]
+                    # Extract columns' names and order them to have # at the beginning
+                    # columns = [desc[0] for desc in result_description]
+                    columns = []
+                    for column in result_columns:
+                        if column.startswith("#"):
+                            # Add at the beginning of the llist
+                            columns.insert(0, column)
+                        else:
+                            # Add at the end of the list
+                            columns.append(column)
 
                     # Return columns as list
                     return columns
@@ -2140,6 +2191,7 @@ class Database:
         compresslevel: int = 6,
         export_header: bool = True,
         sample_list: list = None,
+        force_cast_as_flat: bool = False,
     ) -> bool:
         """
         The `export` function exports data from a database to a specified output format, compresses it
@@ -2227,6 +2279,12 @@ class Database:
         this parameter will be included in the output file. If not provided, the function will determine
         the samples to include based on the data
         :type sample_list: list
+        :param force_cast_as_flat: Only for Parquet format. The `force_cast_as_flat` parameter is a boolean
+        flag that indicates whether to force the export of the data as a flat file format, even if the data
+        is in a nested format. If `force_cast_as_flat` is set to `True`, the data will be exported as a
+        flat file, regardless of its original format. If set to `False`, the data will be exported in its
+        original format. By default, it is set to `False`, defaults to False
+        :type force_cast_as_flat: bool (optional)
         :return: The `export` function returns a boolean value indicating whether the export was
         successful or not.
         """
@@ -2560,7 +2618,8 @@ class Database:
                     """
 
                 # Cast query column to be able to export into a flat file
-                query = cast_columns_query(query=query, conn=self.conn)
+                if output_type not in ["parquet"] or force_cast_as_flat:
+                    query = cast_columns_query(query=query, conn=self.conn)
 
                 # Test empty query
                 df = self.conn.execute(query).fetch_record_batch(1)