+++ RELEASE_NOTES.md
@@ -38,6 +38,7 @@ and transcripts mapping. New useful tools and plugins.
 - Improve stats generation (speedup, annotations stats, custom queries)
 - Improve genomes download with URL or local file as original FASTA
 - Improve dbNSFP download with URL or local file
+- Improve export in Parquet format by keeping column types
 
 ### Fixes
 
@@ -45,6 +46,7 @@ and transcripts mapping. New useful tools and plugins.
 - Fix dnSNP download and index generated vcf
 - Fix explode_infos parameter
 - Fix empty VCF fails
+- Fix header generation
 
 ## 0.12.1.1